% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spa_algorithm_demo_app.R
\name{spa_algorithm_demo_app}
\alias{spa_algorithm_demo_app}
\title{Runs the SPA Algorithm Demo web application}
\usage{
spa_algorithm_demo_app()
}
\description{
Runs the SPA Algorithm Demo web application
}
