% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spa_algorithm_demo_server.R
\name{spa_algorithm_demo_server}
\alias{spa_algorithm_demo_server}
\title{Server of the SPA Algorithm Demo web application}
\usage{
spa_algorithm_demo_server(input, output, session)
}
\description{
Server of the SPA Algorithm Demo web application
}
