% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spa_algorithm_demo_ui.R
\docType{data}
\name{spa_algorithm_demo_ui}
\alias{spa_algorithm_demo_ui}
\title{UI of the SPA Algorithm Demo web application}
\format{
An object of class \code{bslib_page} (inherits from \code{shiny.tag.list}, \code{list}) of length 5.
}
\usage{
spa_algorithm_demo_ui
}
\description{
UI of the SPA Algorithm Demo web application
}
\keyword{datasets}
